Fixed clash with systems headers
This cause very subtle bugs in different components.  Since endian.h
clashed with the one provided by the system, endian detection for system
function did not work properly when our endian.h was included instead of
the one provided by the system.

The include order was not the same on all systems.  This may justify
further investigation.  Because of this difference, the code worked on
one system that was used for development, but not on others.