Reuse objects from rmn to build rmnshared (same for -ompi)
It turns out that object libraries are unnecessary.  Instead of

    add_library(rmn-obj   OBJECT ${sources})
    add_library(rmn       STATIC $<TARGET_OBJECTS:rmn-obj>)
    add_library(rmnshared SHARED $<TARGET_OBJECTS:rmn-obj>)

we can just do

    add_library(rmn       STATIC ${sources})
    add_library(rmnshared SHARED $<TARGET_OBJECTS:rmn>)

to use the objects from `rmn` to build `rmnshared`.

I tested it with rmn used through add_subdirectory() in GEM and it
cmake doesn't complain about cyclic dependencies.

But we have to make sure that App always depends on the STATIC version
of librmn because that is the only case where dependency cycles are
allowed.