Adjust interpv templates for vscode compatibility
The mock-template code that uses the C preprocessor to
define multiple versions of Interp/Extrap functions is
incompatible with Fortran linting, crashing the linter
(see https://github.com/fortran-lang/fortls/issues/236)
because it tries to process the Fortran syntax without
invoking the preprocessor.

This code adjusts the templates to hold just the subroutine
body, leaving the `end subroutine` statement for the
enclosing Fortran file.  This change fixes the linter crash,
although it still doesn't properly understand the
subroutines.