CI: Remove triggers for fst-tools and vgrid
tdpack will trigger fst-tools, which will trigger vgrid.

tdpack and fst-tools (editfst) are needed for the compilation and tests of
vgrid. The use of concurrent gitlab runners brings a problem, as vgrid CI
starts even if those projects are not installed yet. So we are cascading CI
from tdpack to fst-tools to vgrid.