Few improvements ...
Add checking for SIGINT in case we're doing 1800 files.

Use malloc instead of calloc, I don't know but I would think it's faster
since calloc zeroes out the memory and malloc doesn't.

Use strncpy instead of my own strncpytrm since I assume it's faster.

It's hard to tell because the variation in time between two runs of the
same code is so large that it's hard to tell if my code changes have any
effect.